create summarize card effect
