+++ client/src/battle/cardEffect/attack.ts
@@ -1,0 +1,14 @@
+import { CardEffectProps } from '../../types/battle/cardEffect'
+import { oneAttack } from '../../common/cardEffect'
+
+export const strike = (props: CardEffectProps): void => {
+  oneAttack(props)
+}
+
+export const scorpion = (props: CardEffectProps): void => {
+  oneAttack(props)
+}
+
+export const moleclaw = (props: CardEffectProps): void => {
+  oneAttack(props)
+}
