+++ client/src/battle/cardEffect/skill.ts
@@ -1,0 +1,6 @@
+import { CardEffectProps } from '../../types/battle/cardEffect'
+import { guardSkill } from '../../common/cardEffect'
+
+export const protection = (props: CardEffectProps): void => {
+  guardSkill(props)
+}
